Update patch set 1
Patch Set 1:

I've been looking at LPA support, and please correct me if I'm wrong, but it appears the qcom driver is the only piece with LPA coded.  Everything else (framework, system, etc) do not.  If that's correct, is someone working on this (either here or caf)?  If not, I'll try and get something together.

Patch-set: 1