Update patch set 1
Patch Set 1:

The big difference between Q and your devices is that Q uses  the MIPI DSI video interface instead of LCDC. LCDC path has not been tested extensively. If the issue is in the new video code, I suspect its in the LCDC code because we are rock solid on Quincy (and D2, the code is similar).

I don't currently have any Sky/Herc devices to test with so I can only make guesses.

Patch-set: 1