Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Steve, looks like you may be moving us to inheriting from qcom-common.  If that is happening in the near term, I can abandon this change.  If not, this is just a simple port of your latest power HAL for use with the CPU boost hint commits.

Patch-set: 1