Update patch set 1
Patch Set 1: Verified; Looks good to me, but someone else must approve

CM9 and CM10 map internal sdcard to /sdcard and external sdcard to /external_sd.  In Jelly Bean, the storage configuration changed to use /storage for all the media mounts.  But, other devices have created symlinks to keep backwards compatibility.

It seems that the CM10.1 init scripts are creating the /sdcard symlink, but not the /external_sd symlink.  Instead, the symlink to /mnt/extSdCard, which has never been used on skyrocket, hercules or quincy to my knowledge.

I propose we change the /mnt/extSdCard symlink to /external_sd.

Patch-set: 1