Update patch set 3
Patch Set 3:

Yea, Nick the fix is to change the storage config in recovery.fstab. But that undoes some of the policy changes we made in cm-10.1. I need to respond to the JIRA. I think we are going to have to make the recovery.fstab change in cm-10 and reissue recovery images to rommanager.

Patch-set: 3