Update patch set 1
Patch Set 1:

Steve, to reproduce:

1. Install the SR recovery from http://www.clockworkmod.com/rommanager.  Or TWRP or seemly any cm-10.1 based recovery.

2. Wipe /cache, wipe /data.  Install cm-10.2 zip.

3. Boot. On this first boot, kmsg shows e2fsck finds all clean filesystems.

4. Reboot.  On this second boot, kmsg shows e2fsck finds errors on p25 (/data) and usually also finds errors on p26 (/cache).

5. Reboot again.  On this third and any subsequent boots thereafter, kmsg shows e2fsck finds all clean filesystems.  

Note: This problem may have been there all along--but it is more noticeable in cm-10.2 because the e2fsck is run before the CM bootanim runs. The user just sees a blank screen for however long e2fsck takes (the time is variable of course).  The user doesn't wait long enough and he battery pulls thinking he has a soft brick. ..and so on.

Note2: I have not looked for the remount,ro in the kmsg from the 1st boot.  That would be interesting to verify if maybe it doesn't happen on the 1st boot for some reason.

Important clue: Flashing a recovery from cm-10.2 and repeating the above steps results in no corruption at all.  e2fsck is always clean.  This happens with or without 48038 applied to the recovery build and/or the rom build.

Please take these statements with a grain of salt. This is my experience and seems to be confirmed by a few of the other celox guys.  Thanks for looking into this.

Patch-set: 1