Update patch set 1
Patch Set 1:

Dave is there a good reason for having it here? It's causing quite a bit of pain :(

We should just set the composition mode in the init.recovery.rc if it's really needed. When I'm done with the vold integration, this is obsolete because we can just use the "volume share" command which handles it automatically.

Patch-set: 1