Update patch set 1
Patch Set 1:

Ethan/Steve, does it make sense to go down the road of having a camera wrapper for these devices? We always seems to get stuck waiting on our camera workarounds to get ported to the new version of CM. My understanding is that this approach would give us some abstraction and isolation away from the changing implementation of the ever-changing camera java code.

Currently, the wrapper does nothing but prove that the wrapper can work. I have not ported any of our workarounds to the wrapper yet.

Patch-set: 1