Update patch set 2
Patch Set 2:

Hey guys any update on this?  We still have a few issues to sort out on the platform and it seems like everything has stalled ...

1) Camcorder lag on recording start.
2) Panorama photos are out of focus.
3) Audio changes from headphones to external speaker when you tap 'home' while listening to music in Play, Pandora, etc.

Should we discuss what the path forward between these bugs, the pending changes and NEXT branch on the kernel?  Seems like a lot of moving parts to me.

Patch-set: 2