Update patch set 12
Patch Set 12:

I got a java.lang.RuntimeException on Google Camera when trying to change to Video mode which causes a "Can't connect to the camera" message. This is on the 05/06 nightly which according to cmxlog, has this merged on it. Preview is good on stock Camera, but 1080p is corrupt and panorama preview is blurry.

Patch-set: 12
Reviewer: Andrew Arbour <9560@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0