Update patch set 12
Patch Set 12:

Derp.  That's a good call out, all of my validation was on the -next branch of the kernel with media-caf enabled.  I didn't validate this on the current branch. Regardles, we're working to get that all merged very soon - so it should be a moot point within a week.  Thanks for the feedback Andrew.

Patch-set: 12