Update patch set 3
Patch Set 3:

Fair enough, you are getting the result of some reckless hacking by others. Seems a lot of what you just clearly explained could have been put where it belongs, in the commit message.
As for your random determination as to ones ability to "be on gerrit", that is silly. So is the notion that Dev's don't compile AOSP roms/kernels "inline". 
You have to consider that somebody that isn't in on the conversation or happens to find this 2 years later won't have the first clue what your doing/talking about. This happens all the time and causes hours of crap tracing laziness.

Patch-set: 3