Update patch set 6
Patch Set 6:

OK, I know for sure that for Skyrocket and Hercules I have never considered media-caf. Haven't even seen it work. So this is news to me. The odd time that the media folder might be missing and it tries to use media-caf the 300 or so compile errors usually starting with mm-video always remind me that there is something wrong with media folder. lol 
I can just revert this locally and go back and consider what your doing here so I can go over my kernel.
Thanks.

Patch-set: 6