Update patch set 6
Patch Set 6:

I'm compiling for the Skyrocket, but I'm not using the official CM kernel. I'm using vidc drivers that are very close to the ones in the CM kernel and media-caf compiles and works fine for me.

The 5-second delay fix for the camcorder is not in the media HAL, it's in the kernel drivers. This CAF commit fixed the 5-second delay: https://github.com/sultanxda/sultan-kernel-pyramid-pure/commit/2c44f2e1b727a5222721602412f1aff9e81ac0e1

Patch-set: 6