Update patch set 6
Patch Set 6:

Sultan	
hey, thanks for that. Can I ask why? The 2 media folders diverged so long ago. It's a tiny bit more reasonable if your doing a private build but still a strange choice seeing as it's never been done before. Is the non caf version being phased out? 
Why CM is deciding to is the strangest as it was them that decided to fork them to begin with. I would love to know that one. By having to add the media-caf rule in msm8660-common makes it clear if somebody was questioning my thoughts.

Patch-set: 6